moves all Options into a single package, since each may contain both stable and unstable arguments
