+++ besu/src/main/java/org/hyperledger/besu/cli/options/EthstatsOptions.java
@@ -12,9 +12,8 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.cli.options.stable;
+package org.hyperledger.besu.cli.options;
 
-import org.hyperledger.besu.cli.options.CLIOptions;
 import org.hyperledger.besu.ethstats.util.EthStatsConnectOptions;
 
 import java.nio.file.Path;
