+++ besu/src/test/java/org/hyperledger/besu/cli/options/MetricsOptionsTest.java
@@ -14,7 +14,6 @@
  */
 package org.hyperledger.besu.cli.options;
 
-import org.hyperledger.besu.cli.options.stable.MetricsOptions;
 import org.hyperledger.besu.metrics.prometheus.MetricsConfiguration;
 
 import org.junit.jupiter.api.extension.ExtendWith;
